dataloader/pyproject.toml
[build-system]
requires = ["scikit-build-core>=0.9.4", "pybind11"]
build-backend = "scikit_build_core.build"

[project]
name = "native_dataloader"
version = "0.0.1"
description="A custom native data loader for machine learning projects."
readme = "README.md"
authors = [
    { name = "Erik Stern", email = "erik.stern@outlook.de" },
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[tool.scikit-build]
wheel.expand-macos-universal-tags = true
minimum-version = "build-system.requires"
install.strip = false

[tool.pytest.ini_options]
minversion = "8.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
xfail_strict = true
log_cli_level = "INFO"
filterwarnings = [
    "error",
    "ignore::pytest.PytestCacheWarning",
]
testpaths = ["tests"]

[tool.cibuildwheel]
build-frontend = "build[uv]"
test-command = "pytest {project}/tests"
test-extras = ["test"]

[tool.cibuildwheel.pyodide]
build-frontend = {name = "build", args = ["--exports", "whole_archive"]}

[tool.setuptools.packages.find]
where = ["native_dataloader"]

[dependency-groups]
test = [
    "jax[cuda]>=0.4.30",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "pyperclip>=1.11.0",
    "pytest>=8.4.1",
    "pytest-benchmark>=5.2.3",
    "requests>=2.32.5",
]
